requirements.txt
python-dotenv
langchain
langchain_groq
gradio
sentence-transformers
chromadb
